+++ Source/SwiftLintFramework/Rules/LineLengthRule.swift
@@ -14,11 +14,8 @@ public struct LineLengthRule: ParameterizedRule {
     public let identifier = "line_length"
 
     public let parameters = [
-        RuleParameter(severity: .VeryLow, value: 100),
-        RuleParameter(severity: .Low, value: 120),
-        RuleParameter(severity: .Medium, value: 150),
-        RuleParameter(severity: .High, value: 200),
-        RuleParameter(severity: .VeryHigh, value: 250)
+        RuleParameter(severity: .Warning, value: 100),
+        RuleParameter(severity: .Error, value: 200)
     ]
 
     public func validateFile(file: File) -> [StyleViolation] {